refactor : #22 사용하지 않는 주석 삭제, 파일 위치 이동
