--- src/main/java/org/farm/fireflyserver/domain/care/CareAnswer.java
@@ -1,5 +1,0 @@
-package org.farm.fireflyserver.domain.care;
-
-public enum CareAnswer {
-    NORMAL, WARN, EMER
-}
