+++ src/main/java/org/farm/fireflyserver/domain/care/persistence/entity/Care.java
@@ -3,13 +3,10 @@ package org.farm.fireflyserver.domain.care.persistence.entity;
 import jakarta.persistence.*;
 import lombok.*;
 import org.farm.fireflyserver.domain.account.persistence.entity.Account;
-import org.farm.fireflyserver.domain.care.Result;
-import org.farm.fireflyserver.domain.care.Type;
 import org.farm.fireflyserver.domain.senior.persistence.entity.Senior;
 
 
 import org.farm.fireflyserver.common.util.BaseCreatedTimeEntity;
-import org.farm.fireflyserver.domain.senior.persistence.entity.Senior;
 import org.hibernate.annotations.Comment;
 
 import java.time.LocalDateTime;
