+++ src/main/java/org/farm/fireflyserver/domain/care/service/CareServiceImpl.java
@@ -6,8 +6,8 @@ import org.farm.fireflyserver.common.exception.EntityNotFoundException;
 import org.farm.fireflyserver.common.response.ErrorCode;
 import org.farm.fireflyserver.domain.account.persistence.AccountRepository;
 import org.farm.fireflyserver.domain.account.persistence.entity.Account;
-import org.farm.fireflyserver.domain.care.Result;
-import org.farm.fireflyserver.domain.care.Type;
+import org.farm.fireflyserver.domain.care.persistence.entity.Result;
+import org.farm.fireflyserver.domain.care.persistence.entity.Type;
 import org.farm.fireflyserver.domain.care.mapper.CareMapper;
 import org.farm.fireflyserver.domain.care.persistence.AbsentResultRepository;
 import org.farm.fireflyserver.domain.care.persistence.CareRepository;
